pyproject.toml
[tool.poetry]
name = "hanabi_live_deal_scores_retriever"
version = "0.1.0"
description = ""
authors = ["Aaron Liblong <liblonga@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
websocket-client = "^0.57"
requests = "^2.23"
python-dotenv = "^0"

[tool.poetry.scripts]
main = "hanabi_live_deal_scores_retriever.main:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
